requirements.txt
flask
pymongo
flask-cors
python-dotenv